splited gui into several classes
